Se cargan usuarios, Peliculas sigue sin cargar. A espera de más pruebas

+++ src/main/java/um/edu/uy/entities/Evaluacion.java
@@ -5,10 +5,10 @@ import java.time.LocalDate;
 public class Evaluacion {
     private Pelicula pelicula;
     private Usuario user;
-    private int clasificacion;
+    private float clasificacion;
     private LocalDate fecha;
 
-    public Evaluacion (Pelicula pelicula,Usuario user, int clasificacion, LocalDate fecha){
+    public Evaluacion (Pelicula pelicula,Usuario user, float clasificacion, LocalDate fecha){
         this.pelicula = pelicula;
         this.user = new Usuario(user.getId());
         this.clasificacion = clasificacion;
@@ -27,11 +27,11 @@ public class Evaluacion {
         this.user = user;
     }
 
-    public int getClasificacion() {
+    public float getClasificacion() {
         return clasificacion;
     }
 
-    public void setClasificacion(int clasificacion) {
+    public void setClasificacion(float clasificacion) {
         this.clasificacion = clasificacion;
     }
